feat(xmlhttprequest): helper method to set responseType to JSON
Setting the response type can fail on user agents which are implementing support for more recent
editions of the XHR standard (http://xhr.spec.whatwg.org), but fail to support all valid values
for certain properties.

In the case of responseType, relatively recent versions of webkit will fail if the responseType
is 'json', and this is a problem for Safari users, among others.

The main benefit of using responseType is allowing the browser to deserialize JSON, rather than
doing it manually.