feat(data.js): add Merge routine
This is a helper, essentially lodash/underscore's merge. I opt not to depend on lodash or underscore,
as not all applications attempt will want to use them.

The routine performs a shallow copy of the right operands own properties into the left operand. However,
if the left operand has the same property, and it is an object, and the right operand's property is also
an object, then another merge occurs for those two properties.