feat(angular-inbox): build AngularJS version of inbox SDK
This Angular version is nearly identical to vanilla inbox.js, with the exception that it will use
$q rather than native promises, in order to tie directly into the AngularJS life-cycle and update
DOM bindings when responses arrive.

This is a work in progress, as there are some useful additions which I would like to add:

1. Provide a better XMLHttpRequest mock --- using Angular's XHR stuff is not entirely desirable, due
to suffering from a variety of bugs and quirks, and potential difficulty making the API simple.
However, requiring you to use $scope.$digest() to flush responded promises, as well as sinon's mock
XHR, is a bit of a hassle in terms of testability, so it would be good to improve this.

2. Don't check angular.js and angular-mocks.js into the tree. I would like to have a CI server testing
against multiple releases of Angular in the future. Work has been done on writing a Karma plugin to
fetch libraries from bower before running tests, however that work is blocked on a change to Bower,
which currently sucks at programmatically fetching specific versions when other versions are already
checked out.

Beyond this, the rest of the work is really just adding some additional sweetness to the main library,
in order to make interacting with the webservice that much better.