feat(threads): InboxThread#sync() allows thread to be updated from server
Similar to the changes to `InboxAPI#namespaces()` and `InboxNamespace#threads()`, the new method
allows a single Thread to be updated by making a new request to the server, and merging properties
from a successful response into the existing object.

Currently, this will only replace or add new properties, and will never remove old properties from
an existing object. However I believe it will be desirable to have the capacity to delete old items
which are not included in the pulled response. The ability to accomplish this should be forthcoming.