feat(schema): add concept of Schemas to assist in updating existing objects with new data
Schemas allow for there to be a limited set of properties which are processed when new data
arrives from the server, in a limited capacity.

While still naive, with new tests needed, the Schema system allows for a somewhat better experience
when working with objects shared across an entire application.